requirements.txt
paho-mqtt>=1.0
numpy>=1.5.0